+++ src/vgr/builtins/files.py
@@ -1,10 +1,13 @@
 """
-Functions for working with file and dir names.
+Functions for working with files and directories
 """
 
+from pathlib import Path
 from re import Pattern
 from typing import Any
+import grp
 import os
+import pwd
 
 from .type import poly_type
 
@@ -161,6 +164,134 @@ RemoveFile("a_dir") → [False, "[Errno 1] Operation not permitted: 'a_dir'"]
         except OSError as e:
             return [False, str(e)]
     raise ValueError(f'RemoveFile on {poly_type(path)!r} not possible')
+
+def poly_get_file_info(*args) -> Any:
+    """
+**Retrieve information about one or more files**
+
+* *path*.GetFileInfo()
+* GetFileInfo(*path*&hellip;)
+
+Returns one or more, depending upon argument provided, dictionaries
+with information about the requested files.
+All files *must* be relative to the current directory.
+
+The *path* arguments are expressions that, eventually, should resolve to a string.
+Non-string values are ignored, while lists and dictionaries are traversed for
+file paths. In the case of dictionaries, a dictionary using the keys of the
+input is returned, the associated values contain the file information.
+
+
+The contents the information varies depending upon the contents of the
+`status` and `type` attributes.
+
+When the file exists-
+
+```vgr
+{
+    "status": "found",
+    "path": "/path/to/file.txt",
+    "name": "file.txt",
+    "type": "file",
+    "size": 12345,
+    "modified": 1754500000,
+    "created": 1754400000,
+    "is_readable": True,
+    "is_writable": False,
+    "is_executable": False,
+    "owner": "user",
+    "group": "staff"
+}
+```
+
+When `type` is `dir`, the `is_executable` attribute is replaced by
+`is_searchable` and `size` is omitted.
+The availability of creation and modification times, along with `owner` and `group`,
+will vary upon operating systems.
+
+When the file does not exist-
+
+```vgr
+{
+    "status": "not_found",
+    "path": "/path/to/file.txt",
+    "name": "file.txt",
+}
+```
+
+When an error has occurred-
+
+```vgr
+{
+    "status": "error",
+    "path": "/path/to/file.txt",
+    "name": "file.txt",
+    "error": "PermissionError",
+    "message": "Permission denied",
+}
+```
+
+"""
+    def _get_owner(stat):
+        owner = None
+        group = None
+        try:
+            owner = pwd.getpwuid(stat.st_uid).pw_name
+        except KeyError:
+            pass
+        try:
+            group = grp.getgrgid(stat.st_gid).gr_name
+        except KeyError:
+            pass
+        return owner, group
+    def _get_file_info(path: str) -> dict:
+        path: Path = Path(verify_relative_path(path)).resolve()
+        info = {
+            "path": str(path),
+            "name": path.name,
+        }
+        try:
+            if not path.exists():
+                info["status"] = "not_found"
+                return info
+            stat = path.stat()
+            info.update({
+                "status": "found",
+                "type": "dir" if path.is_dir() else "file",
+                "modified": int(stat.st_mtime),
+                "created": int(stat.st_birthtime) if hasattr(stat, "st_birthtime") else int(stat.st_ctime) if os.name == "nt" else None,
+                "is_readable": os.access(path, os.R_OK),
+                "is_writable": os.access(path, os.W_OK),
+                "is_searchable" if path.is_dir() else "is_executable": os.access(path, os.X_OK)
+            })
+            if path.is_file(): info["size"] = stat.st_size
+            owner, group = _get_owner(stat)
+            if owner is not None: info["owner"] = owner
+            if group is not None: info["group"] = group
+            return info
+        except Exception as e:
+            info.update({
+                "status": "error",
+                "error": type(e).__name__,
+                "message": str(e),
+            })
+            return info
+    def _dict_unwrap(v, r):
+        return r[0] if isinstance(v, str) and (isinstance(r, list) and len(r) == 1) else r
+    def _process(value):
+        if isinstance(value, str): return [_get_file_info(value)]
+        if isinstance(value, list):
+            result = []
+            for item in value: result.extend(_process(item))
+            return result
+        if isinstance(value, dict):
+            return [{ k: _dict_unwrap(v, _process(v)) for k, v in value.items() if isinstance(v, (str, list, dict)) }]
+        return []
+    if not args: return None
+    results = []
+    for arg in args: results.extend(_process(arg))
+    # remove superflous array wrapper if they asked for one and got one
+    return results[0] if len(args) == 1 and len(results) == 1 else results
 
 def verify_relative_path(filename: str) -> str:
     """